requirements.txt
# Core dependencies
numpy==1.24.4
beautifulsoup4==4.11.1
scikit-learn==1.3.0
sklearn==0.0
httpx[socks]==0.25.0
python-dotenv==0.20.0

# Utilities
termcolor==1.1.0
toml==0.10.2
validators==0.20.0
tabulate==0.9.0
texttable==1.6.4
treelib==1.7.0
progress==1.6
unipath==1.1

# Data processing
phonenumbers==8.13.22
igraph==0.10.6
yattag==1.15.1
pillow>=9.0.0

# Testing
pytest==7.4.2

# Packaging (only needed for distribution)
pyinstaller==5.13.1
pyinstaller-hooks-contrib==2022.7
